+++ src/hooks/useAddWishlistMutation.ts
@@ -1,4 +1,4 @@
-import { useMutation, UseMutationOptions } from "react-query";
+import { useMutation, UseMutationOptions, useQueryClient } from "react-query";
 import { addWishList } from "../apis/auth";
 
 const useAddWishlistMutation = (
@@ -14,7 +14,21 @@ const useAddWishlistMutation = (
       >
     | undefined,
 ) => {
-  return useMutation(addWishList, options);
+  const queryClient = useQueryClient();
+  return useMutation(addWishList, {
+    onSuccess(res) {
+      if (res.message === "success") {
+        alert("찜 추가 완료");
+        return queryClient.invalidateQueries({
+          queryKey: ["wishlist"],
+        });
+      }
+    },
+    onError(error) {
+      alert("찜 추가 에러: " + error);
+    },
+    ...options,
+  });
 };
 
 export default useAddWishlistMutation;
