+++ src/pages/category/Product.tsx
@@ -18,37 +18,11 @@ const Product = () => {
   const items = useRecoilValue(itemState);
   const token = getCookie("accessToken");
 
-  const { wishlistData, refetch: refetchWishlist } = useWishlistQuery({
-    onSuccess(data) {},
-    onError(error) {
-      alert("찜 리스트 조회 실패: " + error);
-    },
-    enabled: token ? true : false,
-  });
+  const { wishlistData } = useWishlistQuery();
 
-  const addWishListMutation = useAddWishlistMutation({
-    onSuccess(res) {
-      if (res.message === "success") {
-        alert("찜 추가 완료");
-        return refetchWishlist();
-      }
-    },
-    onError(error) {
-      alert("찜 추가 에러: " + error);
-    },
-  });
+  const addWishListMutation = useAddWishlistMutation();
 
-  const deleteWishListMutation = useDeleteWishlistMutation({
-    onSuccess(res) {
-      if (res.message === "success") {
-        alert("찜 삭제 완료");
-        return refetchWishlist();
-      }
-    },
-    onError(error) {
-      alert("찜 삭제 에러: " + error);
-    },
-  });
+  const deleteWishListMutation = useDeleteWishlistMutation();
 
   const toggleWishlist = (productId: number) => {
     if (wishlistData?.find((product) => product.productId === productId)) {
